Split impl fmt::Display to a separate module
The syntax.rs file is getting too large, so I'm splitting the
fmt::Display implementations to a separate module.
