+++ yash-syntax/src/syntax.rs
@@ -77,10 +77,8 @@ use crate::parser::lex::Keyword;
 use crate::parser::lex::Operator;
 use crate::parser::lex::TryFromOperatorError;
 use crate::source::Location;
-use itertools::Itertools;
 use std::cell::OnceCell;
 use std::fmt;
-use std::fmt::Write as _;
 #[cfg(unix)]
 use std::os::unix::io::RawFd;
 use std::rc::Rc;
@@ -235,12 +233,6 @@ impl SpecialParam {
     }
 }
 
-impl fmt::Display for SpecialParam {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.as_char().fmt(f)
-    }
-}
-
 /// Error that occurs when a character cannot be parsed as a special parameter
 ///
 /// This error value is returned by the `TryFrom<char>` and `FromStr`
@@ -364,12 +356,6 @@ impl From<usize> for Param {
     }
 }
 
-impl fmt::Display for Param {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.id.fmt(f)
-    }
-}
-
 // TODO Consider implementing FromStr for Param
 
 /// Flag that specifies how the value is substituted in a [switch](Switch)
@@ -385,19 +371,6 @@ pub enum SwitchType {
     Error,
 }
 
-impl fmt::Display for SwitchType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use SwitchType::*;
-        let c = match self {
-            Alter => '+',
-            Default => '-',
-            Assign => '=',
-            Error => '?',
-        };
-        f.write_char(c)
-    }
-}
-
 /// Condition that triggers a [switch](Switch)
 ///
 /// In the lexical grammar of the shell language, a switch condition is an
@@ -409,16 +382,6 @@ pub enum SwitchCondition {
     /// With a colon, the switch is triggered if the parameter is unset or
     /// empty.
     UnsetOrEmpty,
-}
-
-impl fmt::Display for SwitchCondition {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use SwitchCondition::*;
-        match self {
-            Unset => Ok(()),
-            UnsetOrEmpty => f.write_char(':'),
-        }
-    }
 }
 
 /// Parameter expansion [modifier](Modifier) that conditionally substitutes the
@@ -438,12 +401,6 @@ pub struct Switch {
     pub word: Word,
 }
 
-impl fmt::Display for Switch {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}{}{}", self.condition, self.r#type, self.word)
-    }
-}
-
 impl Unquote for Switch {
     fn write_unquoted<W: fmt::Write>(&self, w: &mut W) -> UnquoteResult {
         write!(w, "{}{}", self.condition, self.r#type)?;
@@ -459,17 +416,6 @@ pub enum TrimSide {
     Prefix,
     /// End of the value
     Suffix,
-}
-
-impl fmt::Display for TrimSide {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use TrimSide::*;
-        let c = match self {
-            Prefix => '#',
-            Suffix => '%',
-        };
-        f.write_char(c)
-    }
 }
 
 /// Flag that specifies pattern matching strategy in a [trim](Trim)
@@ -495,17 +441,6 @@ pub struct Trim {
     pub length: TrimLength,
     /// Pattern to be matched with the expanded value.
     pub pattern: Word,
-}
-
-impl fmt::Display for Trim {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.side.fmt(f)?;
-        match self.length {
-            TrimLength::Shortest => (),
-            TrimLength::Longest => self.side.fmt(f)?,
-        }
-        self.pattern.fmt(f)
-    }
 }
 
 impl Unquote for Trim {
@@ -550,18 +485,6 @@ pub struct BracedParam {
     pub location: Location,
 }
 
-impl fmt::Display for BracedParam {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use Modifier::*;
-        match self.modifier {
-            None => write!(f, "${{{}}}", self.param),
-            Length => write!(f, "${{#{}}}", self.param),
-            Switch(ref switch) => write!(f, "${{{}{}}}", self.param, switch),
-            Trim(ref trim) => write!(f, "${{{}{}}}", self.param, trim),
-        }
-    }
-}
-
 impl Unquote for BracedParam {
     fn write_unquoted<W: fmt::Write>(&self, w: &mut W) -> UnquoteResult {
         use Modifier::*;
@@ -597,15 +520,6 @@ pub enum BackquoteUnit {
     Literal(char),
     /// Backslash-escaped single character
     Backslashed(char),
-}
-
-impl fmt::Display for BackquoteUnit {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            BackquoteUnit::Literal(c) => write!(f, "{c}"),
-            BackquoteUnit::Backslashed(c) => write!(f, "\\{c}"),
-        }
-    }
 }
 
 impl Unquote for BackquoteUnit {
@@ -669,24 +583,6 @@ pub enum TextUnit {
 }
 
 pub use TextUnit::*;
-
-impl fmt::Display for TextUnit {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Literal(c) => write!(f, "{c}"),
-            Backslashed(c) => write!(f, "\\{c}"),
-            RawParam { param, .. } => write!(f, "${param}"),
-            BracedParam(param) => param.fmt(f),
-            CommandSubst { content, .. } => write!(f, "$({content})"),
-            Backquote { content, .. } => {
-                f.write_char('`')?;
-                content.iter().try_for_each(|unit| unit.fmt(f))?;
-                f.write_char('`')
-            }
-            Arith { content, .. } => write!(f, "$(({content}))"),
-        }
-    }
-}
 
 impl Unquote for TextUnit {
     fn write_unquoted<W: fmt::Write>(&self, w: &mut W) -> UnquoteResult {
@@ -754,12 +650,6 @@ impl Text {
     }
 }
 
-impl fmt::Display for Text {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.iter().try_for_each(|unit| unit.fmt(f))
-    }
-}
-
 impl Unquote for Text {
     fn write_unquoted<W: fmt::Write>(&self, w: &mut W) -> UnquoteResult {
         self.0.write_unquoted(w)
@@ -788,17 +678,6 @@ pub enum WordUnit {
 }
 
 pub use WordUnit::*;
-
-impl fmt::Display for WordUnit {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Unquoted(dq) => dq.fmt(f),
-            SingleQuote(s) => write!(f, "'{s}'"),
-            DoubleQuote(content) => write!(f, "\"{content}\""),
-            Tilde(s) => write!(f, "~{s}"),
-        }
-    }
-}
 
 impl Unquote for WordUnit {
     fn write_unquoted<W: fmt::Write>(&self, w: &mut W) -> UnquoteResult {
@@ -845,12 +724,6 @@ pub struct Word {
     pub location: Location,
 }
 
-impl fmt::Display for Word {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.units.iter().try_for_each(|unit| write!(f, "{unit}"))
-    }
-}
-
 impl Unquote for Word {
     fn write_unquoted<W: fmt::Write>(&self, w: &mut W) -> UnquoteResult {
         self.units.write_unquoted(w)
@@ -881,15 +754,6 @@ pub enum Value {
 
 pub use Value::*;
 
-impl fmt::Display for Value {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Scalar(word) => word.fmt(f),
-            Array(words) => write!(f, "({})", words.iter().format(" ")),
-        }
-    }
-}
-
 /// Assignment word
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Assign {
@@ -901,12 +765,6 @@ pub struct Assign {
     pub value: Value,
     /// Location of the assignment word
     pub location: Location,
-}
-
-impl fmt::Display for Assign {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}={}", &self.name, &self.value)
-    }
 }
 
 /// Fallible conversion from a word into an assignment
@@ -959,12 +817,6 @@ impl Fd {
 impl From<RawFd> for Fd {
     fn from(raw_fd: RawFd) -> Fd {
         Fd(raw_fd)
-    }
-}
-
-impl fmt::Display for Fd {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
     }
 }
 
@@ -1032,12 +884,6 @@ impl From<RedirOp> for Operator {
     }
 }
 
-impl fmt::Display for RedirOp {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        Operator::from(*self).fmt(f)
-    }
-}
-
 /// Here-document
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct HereDoc {
@@ -1065,19 +911,6 @@ pub struct HereDoc {
     pub content: OnceCell<Text>,
 }
 
-impl fmt::Display for HereDoc {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(if self.remove_tabs { "<<-" } else { "<<" })?;
-
-        // This space is to disambiguate `<< --` and `<<- -`
-        if let Some(Unquoted(Literal('-'))) = self.delimiter.units.first() {
-            f.write_char(' ')?;
-        }
-
-        write!(f, "{}", self.delimiter)
-    }
-}
-
 /// Part of a redirection that defines the nature of the resulting file descriptor
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum RedirBody {
@@ -1094,15 +927,6 @@ impl RedirBody {
         match self {
             RedirBody::Normal { operand, .. } => operand,
             RedirBody::HereDoc(here_doc) => &here_doc.delimiter,
-        }
-    }
-}
-
-impl fmt::Display for RedirBody {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            RedirBody::Normal { operator, operand } => write!(f, "{operator}{operand}"),
-            RedirBody::HereDoc(h) => write!(f, "{h}"),
         }
     }
 }
@@ -1136,15 +960,6 @@ impl Redir {
             },
             RedirBody::HereDoc { .. } => Fd::STDIN,
         })
-    }
-}
-
-impl fmt::Display for Redir {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if let Some(fd) = self.fd {
-            write!(f, "{fd}")?;
-        }
-        write!(f, "{}", self.body)
     }
 }
 
@@ -1184,40 +999,11 @@ impl SimpleCommand {
     }
 }
 
-impl fmt::Display for SimpleCommand {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let i1 = self.assigns.iter().map(|x| x as &dyn fmt::Display);
-        let i2 = self.words.iter().map(|x| x as &dyn fmt::Display);
-        let i3 = self.redirs.iter().map(|x| x as &dyn fmt::Display);
-
-        if !self.assigns.is_empty() || !self.first_word_is_keyword() {
-            write!(f, "{}", i1.chain(i2).chain(i3).format(" "))
-        } else {
-            // If the simple command starts with an assignment or redirection,
-            // the first word may be a keyword which is treated as a plain word.
-            // In this case, we need to avoid the word being interpreted as a
-            // keyword by printing the assignment or redirection first.
-            write!(f, "{}", i3.chain(i2).format(" "))
-        }
-    }
-}
-
 /// `elif-then` clause
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct ElifThen {
     pub condition: List,
     pub body: List,
-}
-
-impl fmt::Display for ElifThen {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "elif {:#} then ", self.condition)?;
-        if f.alternate() {
-            write!(f, "{:#}", self.body)
-        } else {
-            write!(f, "{}", self.body)
-        }
-    }
 }
 
 /// Symbol that terminates the body of a case branch and determines what to do
@@ -1268,12 +1054,6 @@ impl From<CaseContinuation> for Operator {
     }
 }
 
-impl fmt::Display for CaseContinuation {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        Operator::from(*self).fmt(f)
-    }
-}
-
 /// Branch item of a `case` compound command
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct CaseItem {
@@ -1286,18 +1066,6 @@ pub struct CaseItem {
     pub body: List,
     /// What to do after executing the body of this item
     pub continuation: CaseContinuation,
-}
-
-impl fmt::Display for CaseItem {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "({}) {}{}",
-            self.patterns.iter().format(" | "),
-            self.body,
-            self.continuation,
-        )
-    }
 }
 
 /// Command that contains other commands
@@ -1329,51 +1097,6 @@ pub enum CompoundCommand {
     // TODO [[ ]]
 }
 
-impl fmt::Display for CompoundCommand {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use CompoundCommand::*;
-        match self {
-            Grouping(list) => write!(f, "{{ {list:#} }}"),
-            Subshell { body, .. } => write!(f, "({body})"),
-            For { name, values, body } => {
-                write!(f, "for {name}")?;
-                if let Some(values) = values {
-                    f.write_str(" in")?;
-                    for value in values {
-                        write!(f, " {value}")?;
-                    }
-                    f.write_char(';')?;
-                }
-                write!(f, " do {body:#} done")
-            }
-            While { condition, body } => write!(f, "while {condition:#} do {body:#} done"),
-            Until { condition, body } => write!(f, "until {condition:#} do {body:#} done"),
-            If {
-                condition,
-                body,
-                elifs,
-                r#else,
-            } => {
-                write!(f, "if {condition:#} then {body:#} ")?;
-                for elif in elifs {
-                    write!(f, "{elif:#} ")?;
-                }
-                if let Some(r#else) = r#else {
-                    write!(f, "else {else:#} ")?;
-                }
-                f.write_str("fi")
-            }
-            Case { subject, items } => {
-                write!(f, "case {subject} in ")?;
-                for item in items {
-                    write!(f, "{item} ")?;
-                }
-                f.write_str("esac")
-            }
-        }
-    }
-}
-
 /// Compound command with redirections
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct FullCompoundCommand {
@@ -1381,14 +1104,6 @@ pub struct FullCompoundCommand {
     pub command: CompoundCommand,
     /// Redirections
     pub redirs: Vec<Redir>,
-}
-
-impl fmt::Display for FullCompoundCommand {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let FullCompoundCommand { command, redirs } = self;
-        write!(f, "{command}")?;
-        redirs.iter().try_for_each(|redir| write!(f, " {redir}"))
-    }
 }
 
 /// Function definition command
@@ -1402,15 +1117,6 @@ pub struct FunctionDefinition {
     pub body: Rc<FullCompoundCommand>,
 }
 
-impl fmt::Display for FunctionDefinition {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if self.has_keyword {
-            f.write_str("function ")?;
-        }
-        write!(f, "{}() {}", self.name, self.body)
-    }
-}
-
 /// Element of a pipe sequence
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum Command {
@@ -1420,16 +1126,6 @@ pub enum Command {
     Compound(FullCompoundCommand),
     /// Function definition command
     Function(FunctionDefinition),
-}
-
-impl fmt::Display for Command {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Command::Simple(c) => c.fmt(f),
-            Command::Compound(c) => c.fmt(f),
-            Command::Function(c) => c.fmt(f),
-        }
-    }
 }
 
 /// Commands separated by `|`
@@ -1444,15 +1140,6 @@ pub struct Pipeline {
     pub commands: Vec<Rc<Command>>,
     /// Whether the pipeline begins with a `!`
     pub negation: bool,
-}
-
-impl fmt::Display for Pipeline {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> fmt::Result {
-        if self.negation {
-            write!(f, "! ")?;
-        }
-        write!(f, "{}", self.commands.iter().format(" | "))
-    }
 }
 
 /// Condition that decides if a [Pipeline] in an [and-or list](AndOrList) should be executed
@@ -1484,29 +1171,11 @@ impl From<AndOr> for Operator {
     }
 }
 
-impl fmt::Display for AndOr {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            AndOr::AndThen => write!(f, "&&"),
-            AndOr::OrElse => write!(f, "||"),
-        }
-    }
-}
-
 /// Pipelines separated by `&&` and `||`
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct AndOrList {
     pub first: Pipeline,
     pub rest: Vec<(AndOr, Pipeline)>,
-}
-
-impl fmt::Display for AndOrList {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.first)?;
-        self.rest
-            .iter()
-            .try_for_each(|(c, p)| write!(f, " {c} {p}"))
-    }
 }
 
 /// Element of a [List]
@@ -1521,51 +1190,14 @@ pub struct Item {
     pub async_flag: Option<Location>,
 }
 
-/// Allows conversion from Item to String.
-///
-/// By default, the `;` terminator is omitted from the formatted string.
-/// When the alternate flag is specified as in `{:#}`, the result is always
-/// terminated by either `;` or `&`.
-impl fmt::Display for Item {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.and_or)?;
-        if self.async_flag.is_some() {
-            write!(f, "&")
-        } else if f.alternate() {
-            write!(f, ";")
-        } else {
-            Ok(())
-        }
-    }
-}
-
 /// Sequence of [and-or lists](AndOrList) separated by `;` or `&`
 ///
 /// It depends on context whether an empty list is a valid syntax.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct List(pub Vec<Item>);
 
-/// Allows conversion from List to String.
-///
-/// By default, the last `;` terminator is omitted from the formatted string.
-/// When the alternate flag is specified as in `{:#}`, the result is always
-/// terminated by either `;` or `&`.
-impl fmt::Display for List {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if let Some((last, others)) = self.0.split_last() {
-            for item in others {
-                write!(f, "{item:#} ")?;
-            }
-            if f.alternate() {
-                write!(f, "{last:#}")
-            } else {
-                write!(f, "{last}")
-            }
-        } else {
-            Ok(())
-        }
-    }
-}
+/// Implementations of [std::fmt::Display] for the shell language syntax types
+mod impl_display;
 
 #[allow(clippy::bool_assert_comparison)]
 #[cfg(test)]
@@ -1591,37 +1223,6 @@ mod tests {
     }
 
     #[test]
-    fn switch_display() {
-        let switch = Switch {
-            r#type: SwitchType::Alter,
-            condition: SwitchCondition::Unset,
-            word: "".parse().unwrap(),
-        };
-        assert_eq!(switch.to_string(), "+");
-
-        let switch = Switch {
-            r#type: SwitchType::Default,
-            condition: SwitchCondition::UnsetOrEmpty,
-            word: "foo".parse().unwrap(),
-        };
-        assert_eq!(switch.to_string(), ":-foo");
-
-        let switch = Switch {
-            r#type: SwitchType::Assign,
-            condition: SwitchCondition::UnsetOrEmpty,
-            word: "bar baz".parse().unwrap(),
-        };
-        assert_eq!(switch.to_string(), ":=bar baz");
-
-        let switch = Switch {
-            r#type: SwitchType::Error,
-            condition: SwitchCondition::Unset,
-            word: "?error".parse().unwrap(),
-        };
-        assert_eq!(switch.to_string(), "??error");
-    }
-
-    #[test]
     fn switch_unquote() {
         let switch = Switch {
             r#type: SwitchType::Default,
@@ -1640,37 +1241,6 @@ mod tests {
         let (unquoted, is_quoted) = switch.unquote();
         assert_eq!(unquoted, "?error");
         assert_eq!(is_quoted, true);
-    }
-
-    #[test]
-    fn trim_display() {
-        let trim = Trim {
-            side: TrimSide::Prefix,
-            length: TrimLength::Shortest,
-            pattern: "foo".parse().unwrap(),
-        };
-        assert_eq!(trim.to_string(), "#foo");
-
-        let trim = Trim {
-            side: TrimSide::Prefix,
-            length: TrimLength::Longest,
-            pattern: "".parse().unwrap(),
-        };
-        assert_eq!(trim.to_string(), "##");
-
-        let trim = Trim {
-            side: TrimSide::Suffix,
-            length: TrimLength::Shortest,
-            pattern: "bar".parse().unwrap(),
-        };
-        assert_eq!(trim.to_string(), "%bar");
-
-        let trim = Trim {
-            side: TrimSide::Suffix,
-            length: TrimLength::Longest,
-            pattern: "*".parse().unwrap(),
-        };
-        assert_eq!(trim.to_string(), "%%*");
     }
 
     #[test]
@@ -1710,44 +1280,6 @@ mod tests {
         let (unquoted, is_quoted) = trim.unquote();
         assert_eq!(unquoted, "%%?");
         assert_eq!(is_quoted, false);
-    }
-
-    #[test]
-    fn braced_param_display() {
-        let param = BracedParam {
-            param: Param::variable("foo"),
-            modifier: Modifier::None,
-            location: Location::dummy(""),
-        };
-        assert_eq!(param.to_string(), "${foo}");
-
-        let param = BracedParam {
-            modifier: Modifier::Length,
-            ..param
-        };
-        assert_eq!(param.to_string(), "${#foo}");
-
-        let switch = Switch {
-            r#type: SwitchType::Assign,
-            condition: SwitchCondition::UnsetOrEmpty,
-            word: "bar baz".parse().unwrap(),
-        };
-        let param = BracedParam {
-            modifier: Modifier::Switch(switch),
-            ..param
-        };
-        assert_eq!(param.to_string(), "${foo:=bar baz}");
-
-        let trim = Trim {
-            side: TrimSide::Suffix,
-            length: TrimLength::Shortest,
-            pattern: "baz' 'bar".parse().unwrap(),
-        };
-        let param = BracedParam {
-            modifier: Modifier::Trim(trim),
-            ..param
-        };
-        assert_eq!(param.to_string(), "${foo%baz' 'bar}");
     }
 
     #[test]
@@ -1797,14 +1329,6 @@ mod tests {
     }
 
     #[test]
-    fn backquote_unit_display() {
-        let literal = BackquoteUnit::Literal('A');
-        assert_eq!(literal.to_string(), "A");
-        let backslashed = BackquoteUnit::Backslashed('X');
-        assert_eq!(backslashed.to_string(), r"\X");
-    }
-
-    #[test]
     fn backquote_unit_unquote() {
         let literal = BackquoteUnit::Literal('A');
         let (unquoted, is_quoted) = literal.unquote();
@@ -1815,43 +1339,6 @@ mod tests {
         let (unquoted, is_quoted) = backslashed.unquote();
         assert_eq!(unquoted, "X");
         assert_eq!(is_quoted, true);
-    }
-
-    #[test]
-    fn text_unit_display() {
-        let literal = Literal('A');
-        assert_eq!(literal.to_string(), "A");
-        let backslashed = Backslashed('X');
-        assert_eq!(backslashed.to_string(), r"\X");
-
-        let raw_param = RawParam {
-            param: Param::variable("PARAM"),
-            location: Location::dummy(""),
-        };
-        assert_eq!(raw_param.to_string(), "$PARAM");
-
-        let command_subst = CommandSubst {
-            content: r"foo\bar".into(),
-            location: Location::dummy(""),
-        };
-        assert_eq!(command_subst.to_string(), r"$(foo\bar)");
-
-        let backquote = Backquote {
-            content: vec![
-                BackquoteUnit::Literal('a'),
-                BackquoteUnit::Backslashed('b'),
-                BackquoteUnit::Backslashed('c'),
-                BackquoteUnit::Literal('d'),
-            ],
-            location: Location::dummy(""),
-        };
-        assert_eq!(backquote.to_string(), r"`a\b\cd`");
-
-        let arith = Arith {
-            content: Text(vec![literal, backslashed, command_subst, backquote]),
-            location: Location::dummy(""),
-        };
-        assert_eq!(arith.to_string(), r"$((A\X$(foo\bar)`a\b\cd`))");
     }
 
     #[test]
@@ -1940,29 +1427,6 @@ mod tests {
     }
 
     #[test]
-    fn word_unit_display() {
-        let unquoted = Unquoted(Literal('A'));
-        assert_eq!(unquoted.to_string(), "A");
-        let unquoted = Unquoted(Backslashed('B'));
-        assert_eq!(unquoted.to_string(), "\\B");
-
-        let single_quote = SingleQuote("".to_string());
-        assert_eq!(single_quote.to_string(), "''");
-        let single_quote = SingleQuote(r#"a"b"c\"#.to_string());
-        assert_eq!(single_quote.to_string(), r#"'a"b"c\'"#);
-
-        let double_quote = DoubleQuote(Text(vec![]));
-        assert_eq!(double_quote.to_string(), "\"\"");
-        let double_quote = DoubleQuote(Text(vec![Literal('A'), Backslashed('B')]));
-        assert_eq!(double_quote.to_string(), "\"A\\B\"");
-
-        let tilde = Tilde("".to_string());
-        assert_eq!(tilde.to_string(), "~");
-        let tilde = Tilde("foo".to_string());
-        assert_eq!(tilde.to_string(), "~foo");
-    }
-
-    #[test]
     fn word_unquote() {
         let mut word = Word::from_str(r#"~a/b\c'd'"e""#).unwrap();
         let (unquoted, is_quoted) = word.unquote();
@@ -2001,43 +1465,6 @@ mod tests {
             ..word
         };
         assert_eq!(word.to_string_if_literal(), None);
-    }
-
-    #[test]
-    fn scalar_display() {
-        let s = Scalar(Word::from_str("my scalar value").unwrap());
-        assert_eq!(s.to_string(), "my scalar value");
-    }
-
-    #[test]
-    fn array_display_empty() {
-        let a = Array(vec![]);
-        assert_eq!(a.to_string(), "()");
-    }
-
-    #[test]
-    fn array_display_one() {
-        let a = Array(vec![Word::from_str("one").unwrap()]);
-        assert_eq!(a.to_string(), "(one)");
-    }
-
-    #[test]
-    fn array_display_many() {
-        let a = Array(vec![
-            Word::from_str("let").unwrap(),
-            Word::from_str("me").unwrap(),
-            Word::from_str("see").unwrap(),
-        ]);
-        assert_eq!(a.to_string(), "(let me see)");
-    }
-
-    #[test]
-    fn assign_display() {
-        let mut a = Assign::from_str("foo=bar").unwrap();
-        assert_eq!(a.to_string(), "foo=bar");
-
-        a.value = Array(vec![]);
-        assert_eq!(a.to_string(), "foo=()");
     }
 
     #[test]
@@ -2111,145 +1538,6 @@ mod tests {
     }
 
     #[test]
-    fn here_doc_display() {
-        let heredoc = HereDoc {
-            delimiter: Word::from_str("END").unwrap(),
-            remove_tabs: true,
-            content: Text::from_str("here").unwrap().into(),
-        };
-        assert_eq!(heredoc.to_string(), "<<-END");
-
-        let heredoc = HereDoc {
-            delimiter: Word::from_str("XXX").unwrap(),
-            remove_tabs: false,
-            content: Text::from_str("there").unwrap().into(),
-        };
-        assert_eq!(heredoc.to_string(), "<<XXX");
-    }
-
-    #[test]
-    fn here_doc_display_disambiguation() {
-        let heredoc = HereDoc {
-            delimiter: Word::from_str("--").unwrap(),
-            remove_tabs: false,
-            content: Text::from_str("here").unwrap().into(),
-        };
-        assert_eq!(heredoc.to_string(), "<< --");
-
-        let heredoc = HereDoc {
-            delimiter: Word::from_str("-").unwrap(),
-            remove_tabs: true,
-            content: Text::from_str("here").unwrap().into(),
-        };
-        assert_eq!(heredoc.to_string(), "<<- -");
-    }
-
-    #[test]
-    fn redir_display() {
-        let heredoc = HereDoc {
-            delimiter: Word::from_str("END").unwrap(),
-            remove_tabs: false,
-            content: Text::from_str("here").unwrap().into(),
-        };
-
-        let redir = Redir {
-            fd: None,
-            body: heredoc.into(),
-        };
-        assert_eq!(redir.to_string(), "<<END");
-        let redir = Redir {
-            fd: Some(Fd(0)),
-            ..redir
-        };
-        assert_eq!(redir.to_string(), "0<<END");
-        let redir = Redir {
-            fd: Some(Fd(9)),
-            ..redir
-        };
-        assert_eq!(redir.to_string(), "9<<END");
-    }
-
-    #[test]
-    fn simple_command_display() {
-        let mut command = SimpleCommand {
-            assigns: vec![],
-            words: vec![],
-            redirs: vec![].into(),
-        };
-        assert_eq!(command.to_string(), "");
-
-        command
-            .assigns
-            .push(Assign::from_str("name=value").unwrap());
-        assert_eq!(command.to_string(), "name=value");
-
-        command
-            .assigns
-            .push(Assign::from_str("hello=world").unwrap());
-        assert_eq!(command.to_string(), "name=value hello=world");
-
-        command.words.push(Word::from_str("echo").unwrap());
-        assert_eq!(command.to_string(), "name=value hello=world echo");
-
-        command.words.push(Word::from_str("foo").unwrap());
-        assert_eq!(command.to_string(), "name=value hello=world echo foo");
-
-        Rc::make_mut(&mut command.redirs).push(Redir {
-            fd: None,
-            body: RedirBody::from(HereDoc {
-                delimiter: Word::from_str("END").unwrap(),
-                remove_tabs: false,
-                content: Text::from_str("").unwrap().into(),
-            }),
-        });
-        assert_eq!(command.to_string(), "name=value hello=world echo foo <<END");
-
-        command.assigns.clear();
-        assert_eq!(command.to_string(), "echo foo <<END");
-
-        command.words.clear();
-        assert_eq!(command.to_string(), "<<END");
-
-        Rc::make_mut(&mut command.redirs).push(Redir {
-            fd: Some(Fd(1)),
-            body: RedirBody::from(HereDoc {
-                delimiter: Word::from_str("here").unwrap(),
-                remove_tabs: true,
-                content: Text::from_str("ignored").unwrap().into(),
-            }),
-        });
-        assert_eq!(command.to_string(), "<<END 1<<-here");
-
-        command.assigns.push(Assign::from_str("foo=bar").unwrap());
-        assert_eq!(command.to_string(), "foo=bar <<END 1<<-here");
-    }
-
-    #[test]
-    fn simple_command_display_with_keyword() {
-        let command = SimpleCommand {
-            assigns: vec![],
-            words: vec!["if".parse().unwrap()],
-            redirs: vec!["<foo".parse().unwrap()].into(),
-        };
-        assert_eq!(command.to_string(), "<foo if");
-    }
-
-    #[test]
-    fn elif_then_display() {
-        let condition: List = "c 1& c 2".parse().unwrap();
-        let body = "b 1& b 2".parse().unwrap();
-        let elif = ElifThen { condition, body };
-        assert_eq!(format!("{elif}"), "elif c 1& c 2; then b 1& b 2");
-        assert_eq!(format!("{elif:#}"), "elif c 1& c 2; then b 1& b 2;");
-
-        let condition: List = "c&".parse().unwrap();
-        let body = "b&".parse().unwrap();
-        let elif = ElifThen { condition, body };
-        assert_eq!(format!("{elif}"), "elif c& then b&");
-        assert_eq!(format!("{elif:#}"), "elif c& then b&");
-    }
-
-    #[test]
     fn case_continuation_conversions() {
         use CaseContinuation::*;
         for cc in &[Break, FallThrough, Continue] {
@@ -2263,265 +1551,10 @@ mod tests {
     }
 
     #[test]
-    fn case_item_display() {
-        let item = CaseItem {
-            patterns: vec!["foo".parse().unwrap()],
-            body: "".parse::<List>().unwrap(),
-            continuation: CaseContinuation::Break,
-        };
-        assert_eq!(item.to_string(), "(foo) ;;");
-
-        let item = CaseItem {
-            patterns: vec!["bar".parse().unwrap()],
-            body: "echo ok".parse::<List>().unwrap(),
-            continuation: CaseContinuation::Break,
-        };
-        assert_eq!(item.to_string(), "(bar) echo ok;;");
-
-        let item = CaseItem {
-            patterns: ["a", "b", "c"].iter().map(|s| s.parse().unwrap()).collect(),
-            body: "foo; bar&".parse::<List>().unwrap(),
-            continuation: CaseContinuation::Break,
-        };
-        assert_eq!(item.to_string(), "(a | b | c) foo; bar&;;");
-
-        let item = CaseItem {
-            patterns: vec!["foo".parse().unwrap()],
-            body: "bar".parse::<List>().unwrap(),
-            continuation: CaseContinuation::FallThrough,
-        };
-        assert_eq!(item.to_string(), "(foo) bar;&");
-    }
-
-    #[test]
-    fn grouping_display() {
-        let list = "foo".parse::<List>().unwrap();
-        let grouping = CompoundCommand::Grouping(list);
-        assert_eq!(grouping.to_string(), "{ foo; }");
-    }
-
-    #[test]
-    fn for_display_without_values() {
-        let name = Word::from_str("foo").unwrap();
-        let values = None;
-        let body = "echo ok".parse::<List>().unwrap();
-        let r#for = CompoundCommand::For { name, values, body };
-        assert_eq!(r#for.to_string(), "for foo do echo ok; done");
-    }
-
-    #[test]
-    fn for_display_with_empty_values() {
-        let name = Word::from_str("foo").unwrap();
-        let values = Some(vec![]);
-        let body = "echo ok".parse::<List>().unwrap();
-        let r#for = CompoundCommand::For { name, values, body };
-        assert_eq!(r#for.to_string(), "for foo in; do echo ok; done");
-    }
-
-    #[test]
-    fn for_display_with_some_values() {
-        let name = Word::from_str("V").unwrap();
-        let values = Some(vec![
-            Word::from_str("a").unwrap(),
-            Word::from_str("b").unwrap(),
-        ]);
-        let body = "one; two&".parse::<List>().unwrap();
-        let r#for = CompoundCommand::For { name, values, body };
-        assert_eq!(r#for.to_string(), "for V in a b; do one; two& done");
-    }
-
-    #[test]
-    fn while_display() {
-        let condition = "true& false".parse::<List>().unwrap();
-        let body = "echo ok".parse::<List>().unwrap();
-        let r#while = CompoundCommand::While { condition, body };
-        assert_eq!(r#while.to_string(), "while true& false; do echo ok; done");
-    }
-
-    #[test]
-    fn until_display() {
-        let condition = "true& false".parse::<List>().unwrap();
-        let body = "echo ok".parse::<List>().unwrap();
-        let until = CompoundCommand::Until { condition, body };
-        assert_eq!(until.to_string(), "until true& false; do echo ok; done");
-    }
-
-    #[test]
-    fn if_display() {
-        let r#if: CompoundCommand = CompoundCommand::If {
-            condition: "c 1; c 2&".parse().unwrap(),
-            body: "b 1; b 2&".parse().unwrap(),
-            elifs: vec![],
-            r#else: None,
-        };
-        assert_eq!(r#if.to_string(), "if c 1; c 2& then b 1; b 2& fi");
-
-        let r#if: CompoundCommand = CompoundCommand::If {
-            condition: "c 1& c 2;".parse().unwrap(),
-            body: "b 1& b 2;".parse().unwrap(),
-            elifs: vec![ElifThen {
-                condition: "c 3&".parse().unwrap(),
-                body: "b 3&".parse().unwrap(),
-            }],
-            r#else: Some("b 4".parse().unwrap()),
-        };
-        assert_eq!(
-            r#if.to_string(),
-            "if c 1& c 2; then b 1& b 2; elif c 3& then b 3& else b 4; fi"
-        );
-
-        let r#if: CompoundCommand = CompoundCommand::If {
-            condition: "true".parse().unwrap(),
-            body: ":".parse().unwrap(),
-            elifs: vec![
-                ElifThen {
-                    condition: "false".parse().unwrap(),
-                    body: "a".parse().unwrap(),
-                },
-                ElifThen {
-                    condition: "echo&".parse().unwrap(),
-                    body: "b&".parse().unwrap(),
-                },
-            ],
-            r#else: None,
-        };
-        assert_eq!(
-            r#if.to_string(),
-            "if true; then :; elif false; then a; elif echo& then b& fi"
-        );
-    }
-
-    #[test]
-    fn case_display() {
-        let subject = "foo".parse().unwrap();
-        let items = Vec::<CaseItem>::new();
-        let case = CompoundCommand::Case { subject, items };
-        assert_eq!(case.to_string(), "case foo in esac");
-
-        let subject = "bar".parse().unwrap();
-        let items = vec!["foo)".parse::<CaseItem>().unwrap()];
-        let case = CompoundCommand::Case { subject, items };
-        assert_eq!(case.to_string(), "case bar in (foo) ;; esac");
-
-        let subject = "baz".parse().unwrap();
-        let items = vec![
-            "1)".parse::<CaseItem>().unwrap(),
-            "(a|b|c) :&".parse().unwrap(),
-        ];
-        let case = CompoundCommand::Case { subject, items };
-        assert_eq!(case.to_string(), "case baz in (1) ;; (a | b | c) :&;; esac");
-    }
-
-    #[test]
-    fn function_definition_display() {
-        let body = FullCompoundCommand {
-            command: "( bar )".parse::<CompoundCommand>().unwrap(),
-            redirs: vec![],
-        };
-        let fd = FunctionDefinition {
-            has_keyword: false,
-            name: Word::from_str("foo").unwrap(),
-            body: Rc::new(body),
-        };
-        assert_eq!(fd.to_string(), "foo() (bar)");
-    }
-
-    #[test]
-    fn pipeline_display() {
-        let mut p = Pipeline {
-            commands: vec![Rc::new("first".parse::<Command>().unwrap())],
-            negation: false,
-        };
-        assert_eq!(p.to_string(), "first");
-
-        p.negation = true;
-        assert_eq!(p.to_string(), "! first");
-
-        p.commands.push(Rc::new("second".parse().unwrap()));
-        assert_eq!(p.to_string(), "! first | second");
-
-        p.commands.push(Rc::new("third".parse().unwrap()));
-        p.negation = false;
-        assert_eq!(p.to_string(), "first | second | third");
-    }
-
-    #[test]
     fn and_or_conversions() {
         for op in &[AndOr::AndThen, AndOr::OrElse] {
             let op2 = AndOr::try_from(Operator::from(*op));
             assert_eq!(op2, Ok(*op));
         }
-    }
-
-    #[test]
-    fn and_or_list_display() {
-        let p = "first".parse::<Pipeline>().unwrap();
-        let mut aol = AndOrList {
-            first: p,
-            rest: vec![],
-        };
-        assert_eq!(aol.to_string(), "first");
-
-        let p = "second".parse().unwrap();
-        aol.rest.push((AndOr::AndThen, p));
-        assert_eq!(aol.to_string(), "first && second");
-
-        let p = "third".parse().unwrap();
-        aol.rest.push((AndOr::OrElse, p));
-        assert_eq!(aol.to_string(), "first && second || third");
-    }
-
-    #[test]
-    fn list_display() {
-        let and_or = "first".parse::<AndOrList>().unwrap();
-        let item = Item {
-            and_or: Rc::new(and_or),
-            async_flag: None,
-        };
-        let mut list = List(vec![item]);
-        assert_eq!(list.to_string(), "first");
-
-        let and_or = "second".parse().unwrap();
-        let item = Item {
-            and_or: Rc::new(and_or),
-            async_flag: Some(Location::dummy("")),
-        };
-        list.0.push(item);
-        assert_eq!(list.to_string(), "first; second&");
-
-        let and_or = "third".parse().unwrap();
-        let item = Item {
-            and_or: Rc::new(and_or),
-            async_flag: None,
-        };
-        list.0.push(item);
-        assert_eq!(list.to_string(), "first; second& third");
-    }
-
-    #[test]
-    fn list_display_alternate() {
-        let and_or = "first".parse::<AndOrList>().unwrap();
-        let item = Item {
-            and_or: Rc::new(and_or),
-            async_flag: None,
-        };
-        let mut list = List(vec![item]);
-        assert_eq!(format!("{list:#}"), "first;");
-
-        let and_or = "second".parse().unwrap();
-        let item = Item {
-            and_or: Rc::new(and_or),
-            async_flag: Some(Location::dummy("")),
-        };
-        list.0.push(item);
-        assert_eq!(format!("{list:#}"), "first; second&");
-
-        let and_or = "third".parse().unwrap();
-        let item = Item {
-            and_or: Rc::new(and_or),
-            async_flag: None,
-        };
-        list.0.push(item);
-        assert_eq!(format!("{list:#}"), "first; second& third;");
     }
 }
